+++ ilinskaProj/src/test/java/login/Login_InvCrd.java
@@ -1,8 +1,8 @@
 package login;
 
 import baseTest.BaseTest;
+import org.junit.Assert;
 import org.junit.Test;
-import org.openqa.selenium.WebDriver;
 
 public class Login_InvCrd  extends BaseTest {
 @Test
@@ -12,11 +12,9 @@ public class Login_InvCrd  extends BaseTest {
         loginPage.enterLoginIn("auto");
         loginPage.enterPasswwordInSign("123");
         loginPage.clickOnButtonSignIn();
-        loginPage.isPopupDisplay("Error Sign in Message displays on the screen");
-        homePage.isSignINPresent();
-        homePage.isButtonSignOutPresent();
-
-
+        loginPage.isPopupDisplay();
+        loginPage.isSignINPresent();
+       checkExpectedResult("Sign out present", homePage.isButtonSignOutPresent(), false);
 
     }
     }
